llm.py: Disable response streaming in ollama
Currently, (by default) ollama tries to stream its response as fast as possible
to the user. However, our Python script did not expect this, leading to a
mangled response parsing on our side.
To alleviate this, I added the option `"stream": False` to the payload of
the ollama generation method.