Add experiments for `description_to_display_name` (#8)
This adds the code for running the experiments on the experiment data for `description_to_display_name`.

Furthermore, it fixes a bug that I encountered when running ollama:

`llm.py`: Disable response streaming in ollama

Currently, (by default) ollama tries to stream its response as fast as possible
to the user. However, our Python script did not expect this, leading to a
mangled response parsing on our side.
To alleviate this, I added the option `"stream": False` to the payload of
the ollama generation method.